Feat: add feature connection setting && update style modal product

+++ frontend/portal/src/components/ChannelComponent/ECommerce/Products/TableOfProduct.tsx
@@ -8,6 +8,7 @@ import type { ColumnsType } from "antd/es/table";
 import React from "react";
 import { FaEdit } from "react-icons/fa";
 import { MdDelete } from "react-icons/md";
+import Swal from "sweetalert2";
 
 const TableOfProduct: React.FC<TableOfProductProps> = ({
   dataInTable,
@@ -38,6 +39,8 @@ const TableOfProduct: React.FC<TableOfProductProps> = ({
       key: "price",
       width: "15%",
       ...getColumnSearchProps("price"),
+      sorter: (a, b) => a.description.length - b.description.length,
+      sortDirections: ["descend", "ascend"] as ("descend" | "ascend")[],
     },
     {
       title: "ข้อมูลสินค้า",
@@ -60,16 +63,30 @@ const TableOfProduct: React.FC<TableOfProductProps> = ({
       render: (_, record: IProductInTable) => (
         <div>
           <button
-            className="py-3 px-3 bg-orange-400 text-white rounded-l-lg hover:bg-[#555]"
+            className="py-4 px-3 bg-orange-400 text-white rounded-l-lg hover:bg-[#888] shadow-sm"
             onClick={() => {
               handleEdit(record);
             }}
           >
             <FaEdit />
           </button>
-          <button className="py-3 px-3 bg-red-400 text-white rounded-r-lg hover:bg-[#555]"
+          <button
+            className="py-4 px-3 bg-red-400 text-white rounded-r-lg hover:bg-[#888] shadow-sm"
             onClick={() => {
-              handleDeleteProduct(record.key)
+              Swal.fire({
+                icon: "warning",
+                title: "คุณกำลังจะลบรายการสินค้า",
+                text: `หากดำเนินการต่อ ${record.name} จะถูกลบออกจากระบบ`,
+                showCancelButton: true,
+                confirmButtonText: "ลบ",
+                cancelButtonColor: "#00000055",
+                confirmButtonColor: "#f87171",
+                cancelButtonText: "ยกเลิก",
+              }).then((result: { isConfirmed: any }) => {
+                if (result.isConfirmed) {
+                  handleDeleteProduct(record.key);
+                }
+              });
             }}
           >
             <MdDelete />
@@ -79,7 +96,19 @@ const TableOfProduct: React.FC<TableOfProductProps> = ({
     },
   ];
 
-  return <Table columns={columns} dataSource={dataInTable} />;
+  return (
+    <Table
+      columns={columns}
+      dataSource={dataInTable}
+      size="middle"
+      pagination={{
+        pageSize: 10,
+        // showSizeChanger: true,
+        // position: ["bottomCenter"],
+        // className: "bg-orange-50 rounded-lg p-2 shadow-sm",
+      }}
+    />
+  );
 };
 
 export default TableOfProduct;
